+++ packages/nimble-components/src/table-column/base/tests/types.spec.ts
@@ -1,6 +1,5 @@
 import type {
-    TableColumnSortOperation,
-    TableRowSelectionState
+    TableColumnSortOperation
 } from '../types';
 
 describe('Table column base type', () => {
@@ -8,11 +7,5 @@ describe('Table column base type', () => {
         // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
         const sortOperation: TableColumnSortOperation = 'hello';
         expect(sortOperation).toEqual('hello');
-    });
-
-    it('TableRowSelectionState fails compile if assigning arbitrary string values', () => {
-        // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
-        const selectionState: TableRowSelectionState = 'hello';
-        expect(selectionState).toEqual('hello');
     });
 });
