+++ packages/nimble-components/src/table/tests/types.spec.ts
@@ -1,4 +1,4 @@
-import type { TableColumnSortDirection } from '../types';
+import type { TableColumnSortDirection, TableRowSelectionMode, TableRowSelectionState } from '../types';
 
 describe('Table type', () => {
     it('TableColumnSortDirection fails compile if assigning arbitrary string values', () => {
@@ -10,6 +10,12 @@ describe('Table type', () => {
     it('TableRowSelectionMode fails compile if assigning arbitrary string values', () => {
         // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
         const selectionMode: TableRowSelectionMode = 'hello';
-        expect(selectionMode).toEqual('hello');
+        expect(selectionMode!).toEqual('hello');
+    });
+
+    it('TableRowSelectionState fails compile if assigning arbitrary string values', () => {
+        // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
+        const selectionState: TableRowSelectionState = 'hello';
+        expect(selectionState).toEqual('hello');
     });
 });
